Add a setzero<bool> GPU kernel.
Used by a variety of tests.  Without this change things seem to work
with nvcc, but with clang the executor dies because it can't find said
kernel.  I'm not sure, but I think the issue is nvcc playing fast and
loose with template types.

With this change, plus

 * hacks to the crosstool to let us compile with clang
 * hacks to clang to support std::complex, and
 * a minor hack to eigen

The tensorflow tests seem to pass with clang to the extent that they
also pass with nvcc.